+++ camera.h
@@ -6,35 +6,35 @@ class camera
 {
   private:
     point origin;
-    double aspect_ratio;
+    float aspect_ratio;
     point upper_left_corner;
     normed_vec rel_z;
     normed_vec rel_x;
     normed_vec rel_y;
-    double viewport_width;
-    double viewport_height;
-    double znear;
-    double zfar;
+    float viewport_width;
+    float viewport_height;
+    float znear;
+    float zfar;
   
   public:
     camera( point orig
           , vec look_at
-          , double vertical_fov_in_deg
-          , double image_ratio
+          , float vertical_fov_in_deg
+          , float image_ratio
           , vec absolute_y
-          , double znear
+          , float znear
           );
 
     camera( point orig
           , vec look_at
-          , double vertical_fov_in_deg
-          , double image_ratio
+          , float vertical_fov_in_deg
+          , float image_ratio
           , vec absolute_y
-          , double znear
-          , double zfar
+          , float znear
+          , float zfar
           );
 
-    ray get_ray(double horiz_factor, double vert_factor) const;
-    double get_znear() const;
-    double get_zfar() const;
+    ray get_ray(float horiz_factor, float vert_factor) const;
+    float get_znear() const;
+    float get_zfar() const;
 };
