+++ materials.h
@@ -36,9 +36,9 @@ class metal : public material
 {
   public:
     color albedo;
-    double roughness;
+    float roughness;
 
-    metal(const color& a, double r) : albedo{a}, roughness{r < 1 ? r : 1} {};
+    metal(const color& a, float r) : albedo{a}, roughness{r < 1 ? r : 1} {};
 
     virtual std::optional<ray> scatter(
         const ray& r
@@ -50,9 +50,9 @@ class metal : public material
 class dielectric : public material
 {
   public:
-    double refraction_index;
+    float refraction_index;
 
-    dielectric(double ri) : refraction_index{ri} {};
+    dielectric(float ri) : refraction_index{ri} {};
 
     virtual std::optional<ray> scatter(
         const ray& r
@@ -61,6 +61,6 @@ class dielectric : public material
       ) const override;
 };
 
-static normed_vec get_direction(bool b, normed_vec dir, normed_vec n, double ref_ratio);
-static double reflectance(double cos, double refraction_index);
-static double fresnel_reflectance(double cos_incidence, double cos_refraction, double refraction_index);
+static normed_vec get_direction(bool b, normed_vec dir, normed_vec n, float ref_ratio);
+static float reflectance(float cos, float refraction_index);
+static float fresnel_reflectance(float cos_incidence, float cos_refraction, float refraction_index);
